+++ Source/Android/app/src/main/java/org/dolphinemu/dolphinemu/features/cheats/model/AbstractCheat.java
@@ -3,12 +3,9 @@
 package org.dolphinemu.dolphinemu.features.cheats.model;
 
 import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
 
-public abstract class AbstractCheat implements Cheat
+public abstract class AbstractCheat extends ReadOnlyCheat
 {
-  private Runnable mChangedCallback = null;
-
   public boolean supportsCode()
   {
     return true;
@@ -43,25 +40,6 @@ public abstract class AbstractCheat implements Cheat
     return result;
   }
 
-  public void setEnabled(boolean enabled)
-  {
-    setEnabledImpl(enabled);
-    onChanged();
-  }
-
-  public void setChangedCallback(@Nullable Runnable callback)
-  {
-    mChangedCallback = callback;
-  }
-
-  protected void onChanged()
-  {
-    if (mChangedCallback != null)
-      mChangedCallback.run();
-  }
-
   protected abstract int trySetImpl(@NonNull String name, @NonNull String creator,
           @NonNull String notes, @NonNull String code);
-
-  protected abstract void setEnabledImpl(boolean enabled);
 }
